apps/verifier/requirements.txt
# asyncio_mqtt==0.12.1
cloudevents
flask
gunicorn
httpx
logfmter
pydantic==1.9.1
aiomqtt
python-ulid
